# perf: Faster builds by combining similar builds, and using env vars