--- a/package.json
+++ b/package.json
@@ -44,7 +44,7 @@
     "rollup-plugin-json": "^4.0.0",
     "rollup-plugin-multi-entry": "^2.1.0",
     "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-terser": "^4.0.2"
+    "rollup-plugin-terser": "^5.1.1"
   },
   "peerDependencies": {
     "rollup": "^1.12.0"