Noticed a few things that we could do to improve build performance while working on the vhs generator update.

> Note: all numbers are based on VHS build times, which is a bit of an exceptional case.

1. Combine builds `module` and `browser` builds
* terser now supports only running on one file in an output. So having the browser output include both `.min.js` and `.js` saves us around ~3 seconds per build.
* We should have been building modules together all along, this saves us ~2s per build
2. Allow builds to be turned on and off via environment vars
* This allows us to only create the `test` bundle when running tests in ci saving ~15s
* We can also skip the test bundle creation when publishing or running `build` saving ~6s